+++ manual/source/version_history.rst
@@ -21,6 +21,7 @@ New Features
 * Ability to specify the world volume material with the `worldMaterial` option.
 * Introduced `minimumRange` and `minimumKineticEnergy` user limits as provided by G4UserLimits.
 * Ability to limit step size in user-specified fields.
+* Ability to control turn number print out with `printFractionTurns`.
 * Magnet yokes now have a general multipolar field.
 
 
@@ -55,7 +56,8 @@ General
   to absolute paths.
 * More efficient file IO in rebdsim improves analysis speed.
 * The world material is now air instead of G4_Galatic.
-* `includeDironMagFields` option is now `yokeFields`.
+* `printModuloFraction` is now `printFractionEvents`.
+* `includeIronMagFields` option is now `yokeFields`.
 
 Bug Fixes
 ---------
@@ -74,6 +76,7 @@ Bug Fixes
   geometry with strong rotations.
 * When no energy was specified in the `userfile` bunch distribution, the central beam energy was
   1000x too low - fixed - missing factory of GeV.
+* Number of turns simulated is now number desired and not number+1.
 * Limits to control maximum step length and tracking time were not attached to magnet yoke geometry.